+++ src/test/failures.test.ts
@@ -7,6 +7,7 @@
 import {suite} from 'uvu';
 import * as assert from 'uvu/assert';
 import {rigTest} from './util/rig-test.js';
+import type {ExitResult} from './util/test-rig.js';
 
 const test = suite<object>();
 
@@ -477,6 +478,140 @@ test(
 
     // `kill` is killed.
     assert.equal((await wireit.exit).code, 1);
+  }),
+);
+
+test(
+  'unexpected input file deletion during fingerprinting',
+  rigTest(async ({rig}) => {
+    // Spam our input file with writes and deletes out-of-band with wireit.
+    let spamming = true;
+    void (async () => {
+      while (spamming) {
+        try {
+          await rig.write('input', Math.random());
+          await rig.delete('input');
+        } catch {
+          // Sometimes we get an EPERM error here on Windows CI. Probably
+          // writing too fast, just sleep a bit.
+          await new Promise((resolve) => setTimeout(resolve, 1000));
+        }
+      }
+    })();
+
+    let finalExit: ExitResult;
+    try {
+      // It could take multiple attempts to hit the race condition.
+      for (let i = 0; i < 100; i++) {
+        const failer = await rig.newCommand();
+        await rig.write({
+          'package.json': {
+            scripts: {
+              main: 'wireit',
+              failer: 'wireit',
+            },
+            wireit: {
+              main: {
+                dependencies: ['failer'],
+              },
+              failer: {
+                command: failer.command,
+                files: ['input'],
+                output: ['output'],
+              },
+            },
+          },
+        });
+        const wireit = rig.exec('npm run main');
+        // If the error occurs, it will happen before invocation.
+        const exitOrInvocation = await Promise.race([
+          wireit.exit,
+          failer.nextInvocation(),
+        ]);
+        if ('code' in exitOrInvocation) {
+          finalExit = exitOrInvocation;
+          break;
+        }
+        await rig.write('output', '1');
+        exitOrInvocation.exit(0);
+        finalExit = await wireit.exit;
+        if (finalExit.code !== 0) {
+          break;
+        }
+      }
+    } finally {
+      spamming = false;
+    }
+
+    assert.equal(finalExit!.code, 1);
+    assert.match(
+      finalExit!.stderr,
+      `[failer] Input file "${rig.resolve('input')}" was deleted unexpectedly.` +
+        ` Is another process writing to the same location?`,
+    );
+  }),
+);
+
+test(
+  'unexpected output file deletion during manifest generation',
+  rigTest(async ({rig}) => {
+    // Spam our output file with writes and deletes out-of-band with wireit.
+    let spamming = true;
+    void (async () => {
+      while (spamming) {
+        try {
+          await rig.write('output', Math.random());
+          await rig.delete('output');
+        } catch {
+          // Sometimes we get an EPERM error here on Windows CI. Probably
+          // writing too fast, just sleep a bit.
+          await new Promise((resolve) => setTimeout(resolve, 1000));
+        }
+      }
+    })();
+
+    let finalExit: ExitResult;
+    try {
+      // It could take multiple attempts to hit the race condition.
+      for (let i = 0; i < 100; i++) {
+        const failer = await rig.newCommand();
+        await rig.write({
+          'package.json': {
+            scripts: {
+              main: 'wireit',
+              failer: 'wireit',
+            },
+            wireit: {
+              main: {
+                dependencies: ['failer'],
+              },
+              failer: {
+                command: failer.command,
+                files: ['input'],
+                output: ['output'],
+              },
+            },
+          },
+        });
+        const wireit = rig.exec('npm run main');
+        const failerInv = await failer.nextInvocation();
+        await rig.write('output', '1');
+        failerInv.exit(0);
+        finalExit = await wireit.exit;
+        if (finalExit.code !== 0) {
+          break;
+        }
+      }
+    } finally {
+      spamming = false;
+    }
+
+    assert.equal(finalExit!.code, 1);
+    assert.match(
+      finalExit!.stderr,
+      `[failer] Output file "${rig.resolve('output')}" was deleted unexpectedly.` +
+        ` Is another process writing to the same location?`,
+    );
   }),
 );